Add topbar pyproject with jupyter releaser config
We set version in top level package.json as well. From now we need to make
sure that version in toplevel package is in sync will versions
of all packages.

Top level pyproject will have a dummy package too. Not sure if it
will be an issue with check-python workflow.